requirements.txt
streamlit
scikit-learn
pandas
numpy
plotly
groq 
wikipedia-api 
duckduckgo-search
python-dotenv
wikipedia-api
